LPS-99657 Class that implements ConfigurationBeanDeclaration should be in package *.settings.definition
